+++ corehq/apps/sms/models.py
@@ -11,7 +11,7 @@ from django.db import models, transaction
 from collections import namedtuple
 from corehq.apps.app_manager.dbaccessors import get_app
 from corehq.apps.app_manager.models import Form
-from corehq.apps.hqwebapp.models import UUIDGeneratorMixin
+from corehq.util.mixin import UUIDGeneratorMixin
 from corehq.apps.users.models import CouchUser
 from casexml.apps.case.models import CommCareCase
 from dimagi.utils.couch.migration import SyncSQLToCouchMixin
